src/video_downloader/requirements.txt
boto3
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pyparsing<3.0.0
